refactor(profile): rename onload wiring + clear residual Focus→Profile (RFC 0a0791c1 Phase 5 T3-PR4)
Mechanical zero-behavior-change symbol rename, PR 4 of 4 (final — clears the
src gate to zero):
- FocusProfileOnloadWiring → ProfileOnloadWiring (file rename + all import paths,
  src + 3 test imports + 1 comment ref)
- residual Focus/Knowledge-profile prose in src comments (SPARQLApi,
  CreateAssetCommand, ExocortexSettings, PluginLocalDataStore,
  PluginSettingsStoreAdapter) → Profile
- exo__FocusProfile → exo__Profile (test fixtures + comments)
- consistency renames of two Focus-named test files → ProfileSwitch.regression /
  ProfileWiring.integration

PRESERVED (intentional, NOT code symbols): legacy persisted-data keys
`activeKnowledgeProfileUid`/`activeFocusProfileUid` in PluginSettingsStoreAdapter
migration test — they assert the adapter ignores retired on-disk slots from older
plugin versions (backward-compat contract; test-only, out of the src gate).

DoD: `git grep -E "FocusProfile|KnowledgeProfile" -- 'packages/**/src/**'` = 0.
Command id hard-switch-focus-profile preserved. check:types + lint(0 err) green;
133 affected tests pass.
