+++ packages/obsidian-plugin/src/infrastructure/adapters/PluginSettingsStoreAdapter.ts
@@ -11,7 +11,7 @@ import type { PluginLocalDataStore } from "./PluginLocalDataStore";
  * Prior to Item #3, switch state lived в `plugin.data.json` (synced
  * across devices via Obsidian Sync). That caused two UX bugs:
  *   1. Profile selection replicated cross-device against user intent
- *      (per CLAUDE.md FocusProfile section «per-device» contract).
+ *      (per CLAUDE.md Profile section «per-device» contract).
  *   2. `_switchInProgress=true` left on a crashed device surfaced on
  *      sibling devices as spurious `recoverIfNeeded` fires.
  *
