+++ packages/obsidian-plugin/src/infrastructure/adapters/ProfileOnloadWiring.ts
